\name{descr_avalitools}
\alias{descr_avalitools}
\title{Correlation tables for avaliation models.}
\description{Describes the numeric variables in the research.}
\usage{
descr_avalitools(df)
}
\arguments{
\item{df}{
df is a data.frame with the variables (not transformed) represented in the columns.
}

}

\value{
A data.frame with descriptive measures.
}
\examples{
df_descr <- descr_avalitools(df)

}