\name{transf_vari_names}
\alias{transf_vari_names}}
\title{Transform Variable Names.}
\description{Creates an string vector with the transformed variable names}
\usage{
transf_vari_names(name, trans)
}
\arguments{

\item{name}{
name is the string vector of the names .
}

\item{trans}{
trans is the string vector of the transformations.
}

}

\value{
A string vector with the transformed names.
}
\examples{
trasf_names <- transf_vari_names(name, trans)
}